separated out components from fileupload file
